chatbot-local-llm/requirements.txt
streamlit
python_dotenv
langchain_community